fix: 모달 사용하는 곳에서 그리도록 수정, 공통모달 컴포넌트 추상화
